WIP: Move header code to out of `app.js`
